deep_research/requirementss.txt
pydantic
ddgs
bs4
requests
rich
python-dotenv
openai-agents
